requirements.txt
accelerate==1.2.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blis==0.7.11
build==1.2.2.post1
cachetools==5.5.0
catalogue==2.0.10
cbor==1.0.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cloudpathlib==0.16.0
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
cryptography==44.0.0
cymem==2.0.10
dashscope==1.20.14
dataclasses-json==0.6.7
datasets==2.19.0
Deprecated==1.2.15
dill==0.3.8
durationpy==0.9
emoji==2.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.9.0.post1
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
FlagEmbedding==1.3.3
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.3.1
google-auth==2.37.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inscriptis==2.5.0
ir_datasets==0.5.9
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kubernetes==31.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.2.4
language_data==1.3.0
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pathlib_abc==0.1.1
pathy==0.11.0
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.14.0
pillow==11.0.0
pipdeptree==2.24.0
posthog==3.7.4
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
Pygments==2.18.0
pypdf==5.1.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-pkuseg==0.0.33
SQLAlchemy==2.0.36
srsly==2.5.0
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.67.1
transformers==4.44.2
trec-car-tools==2.6
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unlzw3==0.2.3
unstructured==0.16.11
unstructured-client==0.28.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
watchfiles==1.0.3
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.7.0/zh_core_web_sm-3.7.0-py3-none-any.whl#sha256=f51075665749e07406d629d1055ce5a68635fae6ab3c34257ee798c62b4fc431
zipp==3.21.0
zlib-state==0.1.9
